Add Linux support
This uses the libSDL3 system library... this is dumb.  SDL refuses to
ship binaries for Linux, and instead want each distro to ship them.

Eventually we should add building SDL as a part of the generate step, or
find some pre-built binaries that works on a lot of distros and ship
that insead.

...but this works for now.